app/requirements.txt
autopep8==2.0.1
boto3==1.26.67
botocore==1.29.67
celery==5.2.7
channels==4.0.0
channels-redis==4.0.0
dj-database-url==1.2.0
django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-filter==22.1
django-import-export==3.0.2
django-redis==5.2.0
django-storages==1.13.2
drf-spectacular==0.25.1
docutils==0.19
email-validator==1.3.1
Faker==16.7.0
flower==1.2.0
gunicorn==20.1.0
locust==2.14.2
markdown==3.4.1
MarkupSafe==2.1.2
packaging==23.0
pandas==1.5.3
pika==1.3.1
pika-stubs==0.1.3
psutil==5.9.4
psycopg2-binary==2.9.5
pyfcm==1.5.4
pylint==2.16.1
pyotp==2.8.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.7
redis==4.5.1
requests==2.28.2
selenium==4.8.0
sendgrid==6.9.7
sentry-sdk==1.15.0
twilio==7.16.3
whitenoise==6.3.0

